SoC and formatting changes
moved ChatGpt to its own component, formatting and styling changes on code for clarity
